+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `set_max_payload_length` method
 - `rssi_wideband` method
 - `last_rx_byte_addr` method
+- `set_hop_period` method
 
 ## [0.1.0] - 2026-03-07